Refactors Combatant classes and fixes bug with BattleMenu
